requirements.txt
Django==2.2.1
django-ckeditor==5.7.1
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.10.0
google-api-python-client==1.4.1
httplib2==0.12.3
oauth2client==4.1.3
Pillow==6.0.0
psycopg2-binary==2.8.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
